th advance search installation msg issue fixed

+++ inc/header-function.php
@@ -236,12 +236,13 @@ $open_shop_menu_open = get_theme_mod('open_shop_mobile_menu_open','left');
 				   </div>
            <div class="below-header-col2">
              <?php  
-             if ( class_exists( 'WooCommerce' ) && class_exists('TH_Advance_Product_Search')){
+             if ( class_exists('TH_Advance_Product_Search')){
                echo do_shortcode('[th-aps]');
-                
-               }else {
+               }elseif ( !class_exists('TH_Advance_Product_Search') && is_user_logged_in()) {
                 $url = admin_url('themes.php?page=open-shop');
                       echo '<a href="'.$url.'" target="_blank" class="plugin-active-msg">'.__('Please install th advance product search plugin','open-shop').'</a>';
+               }else{
+                echo "&nbsp&nbsp";
                }
             ?>
          </div>
@@ -543,7 +544,7 @@ add_action('open_shop_site_preloader','open_shop_preloader');
                      <?php 
                 if ( class_exists( 'WooCommerce' ) && class_exists('TH_Advance_Product_Search')){
                echo do_shortcode('[th-aps]');
-               }else {
+               }elseif ( !class_exists('TH_Advance_Product_Search') && is_user_logged_in()) {
                 $url = admin_url('themes.php?page=open-shop');
                       echo '<a href="'.$url.'" target="_blank" class="plugin-active-msg">'.__('Please install th advance product search plugin','open-shop').'</a>';
                }?>
